Refactoring done as required by the TSE
